Move the XRay analysis scripts into its own subfolder
