+++ analyzer/rootana/scripts/XRayAnalysis/XRay.h
@@ -1,0 +1,10 @@
+// A useful structure that represents the XRay transition
+struct XRay {
+  std::string transition;
+  std::string material;
+  double energy;
+  double intensity;
+  double intensity_error;
+  double efficiency;
+  double efficiency_error;
+};
